Make toast into its own component
