+++ src/app/_layout.tsx
@@ -2,13 +2,7 @@ import { Stack } from 'expo-router';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 import { AuthContextProvider } from '../utils/AuthContext';
-import Toast, {
-  BaseToast,
-  BaseToastProps,
-  ToastConfig,
-} from 'react-native-toast-message';
-import { Icon } from 'react-native-elements';
-import colors from '../styles/colors';
+import ToastComponent from '../components/Toast/Toast';
 
 function StackLayout() {
   return (
@@ -21,26 +15,9 @@ function StackLayout() {
           <Stack.Screen name="auth" options={{ headerShown: false }} />
         </Stack>
       </AuthContextProvider>
-      <Toast config={toastConfig} />
+      <ToastComponent />
     </SafeAreaProvider>
   );
 }
-
-const toastConfig: ToastConfig = {
-  success: (props: BaseToastProps) => (
-    <BaseToast
-      {...props}
-      style={{ borderLeftColor: colors.white }}
-      renderLeadingIcon={() => (
-        <Icon
-          name="check-circle"
-          type="octicon"
-          color={colors.lime}
-          style={{ flex: 1, justifyContent: 'center', marginLeft: 10 }}
-        />
-      )}
-    />
-  ),
-};
 
 export default StackLayout;
