+++ src/components/Toast/styles.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+import colors from '../../styles/colors';
+
+export default StyleSheet.create({
+  icon: {
+    flex: 1,
+    justifyContent: 'center',
+    marginLeft: 10
+  },
+  toast: {
+    borderLeftColor: colors.white
+  }
+})
